TowerInstruct-7B-v0.2/requirements.txt
torch>=2.1.0
transformers==4.36.2
peft==0.7.1
datasets==2.15.0
accelerate==0.25.0
scipy
sentencepiece
sacrebleu
evaluate
wandb
pyyaml
tqdm
numpy
pandas
nltk